pyproject.toml
[tool.poetry]
name = "module13"
version = "0.1.0"
description = ""
authors = ["Dmytro <777fesenko.d@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
pydantic = "^2.5.2"
uvicorn = "^0.24.0.post1"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
libgravatar = "^1.0.4"
python-multipart = "^0.0.6"
asyncpg = "^0.29.0"
bcrypt = "^4.1.1"
alembic = "^1.13.0"
python-dotenv = "^1.0.0"
cloudinary = "^1.37.0"
redis = "^5.0.1"
fastapi-mail = "^1.4.1"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
